ImageMigration

+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/ImageDao.java
@@ -8,10 +8,10 @@ import java.util.Optional;
 import java.util.stream.Stream;
 
 public interface ImageDao {
-    Stream<Image> findByPetId(long id);
-    Integer quantityByPetId(long id);
-    Optional<byte[]> getDataById(long id);
-    Optional<Image> create(long petId, byte[] bytes);
-    void delete(Integer id);
-    void delete(List<Integer> ids);
+    Stream<Image> findByPetId(Long id);
+    Integer quantityByPetId(Long id);
+    Optional<byte[]> getDataById(Long id);
+    Optional<Image> create(Long petId, byte[] bytes);
+    void delete(Long id);
+    void delete(List<Long> ids);
 }
